Get some new attributes when calling hooks
When calling updated and created callback hooks there might be some new data to add to the model.
This is where I felt like a deadend when I was working with "serverSide: false" in datatables which will not reload the data after edit or create and use the response from the server. So I had to add some columns to the returning data to fix it and this parts that I've added might be good for other usages too.